+++ admin/src/main/java/com/example/admin/rest/AdminAuthorizeController.java
@@ -1,8 +1,7 @@
 package com.example.admin.rest;
 
 import com.example.admin.service.AdminAuthorizeApplicationService;
-import com.example.admin.usecases.authorize.LoginCase;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.example.admin.usecase.authorize.LoginCase;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
