+++ admin/src/main/java/com/example/admin/service/AdminApplicationService.java
@@ -1,9 +1,8 @@
 package com.example.admin.service;
 
-import com.example.admin.usecases.authorize.ResetPasswordCase;
+import com.example.admin.usecase.authorize.ResetPasswordCase;
 import com.example.domain.user.model.Operator;
 import com.example.domain.user.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
