+++ admin/src/main/java/com/example/admin/usecase/group/GetGroupsCase.java
@@ -1,38 +1,33 @@
-package com.example.admin.usecases.group;
+package com.example.admin.usecase.group;
 
 import com.example.domain.group.model.Group;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.validation.constraints.NotBlank;
+import java.time.Instant;
 
-public class CreateGroupCase {
-    @Getter
-    @Setter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class Request {
-        @NotBlank(message = "group_request_id")
-        private String groupRequestId;
-    }
-
+public class GetGroupsCase {
     @Getter
     @Setter
     @Builder
     public static class Response {
         private String id;
+
         private String name;
+
         private String description;
+
+        private Instant createdAt;
 
         public static Response from(Group group) {
             return Response.builder()
                     .id(group.getId())
                     .name(group.getName())
                     .description(group.getDescription())
+                    .createdAt(group.getCreatedAt())
                     .build();
         }
     }
+
 }
