+++ admin/src/main/java/com/example/admin/usecase/question/UpdateQuestionStatusCase.java
@@ -1,6 +1,6 @@
-package com.example.admin.usecases.user;
+package com.example.admin.usecase.question;
 
-import com.example.domain.user.model.User;
+import com.example.domain.question.model.Question;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,14 +9,14 @@ import lombok.Setter;
 
 import javax.validation.constraints.NotNull;
 
-public class UpdateUserStatusCase {
+public class UpdateQuestionStatusCase {
     @Getter
     @Setter
     @NoArgsConstructor
     @AllArgsConstructor
     public static class Request {
-        @NotNull(message = "status_required")
-        private User.Status status;
+        @NotNull(message = "question_status_required")
+        private Question.Status status;
     }
 
     @Getter
@@ -24,12 +24,12 @@ public class UpdateUserStatusCase {
     @Builder
     public static class Response {
         private String id;
-        private User.Status status;
+        private Question.Status status;
 
-        public static Response from(User user) {
+        public static Response from(Question question) {
             return Response.builder()
-                    .id(user.getId())
-                    .status(user.getStatus())
+                    .id(question.getId())
+                    .status(question.getStatus())
                     .build();
         }
     }
