+++ admin/src/main/java/com/example/admin/usecase/user/GetUserDetailCase.java
@@ -1,28 +1,13 @@
-package com.example.admin.usecases.user;
+package com.example.admin.usecase.user;
 
 import com.example.domain.user.model.User;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.validation.constraints.NotBlank;
+import java.time.Instant;
 
-public class CreateUserCase {
-    @Getter
-    @Setter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class Request {
-        @NotBlank(message = "name_required")
-        private String name;
-        @NotBlank(message = "email_required")
-        private String email;
-        @NotBlank(message = "password_required")
-        private String password;
-    }
-
+public class GetUserDetailCase {
     @Getter
     @Setter
     @Builder
@@ -30,12 +15,14 @@ public class CreateUserCase {
         private String id;
         private String name;
         private String email;
+        private Instant createdAt;
 
         public static Response from(User user) {
             return Response.builder()
                     .id(user.getId())
                     .name(user.getName())
                     .email(user.getEmail())
+                    .createdAt(user.getCreatedAt())
                     .build();
         }
     }
