+++ admin/src/main/java/com/example/admin/usecase/user/GetUsersCase.java
@@ -1,4 +1,4 @@
-package com.example.admin.usecases.user;
+package com.example.admin.usecase.user;
 
 import com.example.domain.user.model.User;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Setter;
 
 import java.time.Instant;
 
-public class GetUserDetailCase {
+public class GetUsersCase {
     @Getter
     @Setter
     @Builder
@@ -15,6 +15,7 @@ public class GetUserDetailCase {
         private String id;
         private String name;
         private String email;
+        private User.Status status;
         private Instant createdAt;
 
         public static Response from(User user) {
@@ -22,6 +23,7 @@ public class GetUserDetailCase {
                     .id(user.getId())
                     .name(user.getName())
                     .email(user.getEmail())
+                    .status(user.getStatus())
                     .createdAt(user.getCreatedAt())
                     .build();
         }
